Better support for optional timeout
In Python 2, the existing code works because `any int >= None` is True. This won't work on Python 3 (it raises a TypeError), and in any case the exception message with timeout = None is not clear.

These changes are retro-compatible, but allow for better exception messages (I haven't tested actual Python 3 support yet, but this shouldn't hurt).